Add support for downloading modules
This changes allow the app to download modules based on version

I added a commented structure to the config file for the app so that the
structure gets documented but at the same time so that it does not cause
any side effect.

This changes should not be merged before https://github.com/checheza/checheza-core/pull/5
as it uses a new module introduces there.

Signed-off-by: Joao Pedro Silva <jp.silva15@gmail.com>